requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
gcloud==0.18.3
googleapis-common-protos==1.62.0
gunicorn==21.2.0
httplib2==0.22.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
oauth2client==4.1.3
packaging==23.2
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==1.0.0
rsa==4.9
six==1.16.0
Werkzeug==3.0.1
